compose test classes
